pyproject.toml
[project]
name = "ai-fast-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Horbee",email = "n.horox@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "torch (>=2.7.1,<3.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "google-api-python-client (>=2.171.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
